# backend/openapi.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"summary":"Ouvrir une session","tags":["Authentification"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":""}},"summary":"Renouveler les jetons (usage unique)","tags":["Authentification"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Fermer la session et révoquer le jeton de renouvellement","tags":["Authentification"]}},"/api/auth/device-token":{"post":{"operationId":"AuthController_enregistrerJeton","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceTokenDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Enregistrer ou retirer le jeton FCM de cet appareil","tags":["Authentification"]}},"/api/clients":{"post":{"operationId":"ClientsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Clients"]},"get":{"operationId":"ClientsController_findAll","parameters":[{"name":"recherche","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Clients"]}},"/api/clients/{id}":{"get":{"operationId":"ClientsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Clients"]},"patch":{"operationId":"ClientsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClientDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Clients"]},"delete":{"operationId":"ClientsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Clients"]}},"/api/clients/{id}/adresses":{"post":{"operationId":"ClientsController_ajouterAdresse","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAdresseDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Clients"]}},"/api/fournisseurs":{"post":{"operationId":"FournisseursController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFournisseurDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Fournisseurs"]},"get":{"operationId":"FournisseursController_findAll","parameters":[{"name":"recherche","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Fournisseurs"]}},"/api/fournisseurs/{id}":{"get":{"operationId":"FournisseursController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Fournisseurs"]},"patch":{"operationId":"FournisseursController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFournisseurDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Fournisseurs"]},"delete":{"operationId":"FournisseursController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Fournisseurs"]}},"/api/fournisseurs/{id}/produits":{"post":{"operationId":"FournisseursController_ajouterProduit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProduitDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Fournisseurs"]}},"/api/livreurs":{"post":{"operationId":"LivreursController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLivreurDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Livreurs"]},"get":{"operationId":"LivreursController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Livreurs"]}},"/api/livreurs/{id}":{"get":{"operationId":"LivreursController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Livreurs"]},"patch":{"operationId":"LivreursController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLivreurDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Livreurs"]},"delete":{"operationId":"LivreursController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Livreurs"]}},"/api/livreurs/{id}/statut":{"patch":{"operationId":"LivreursController_updateStatut","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatutLivreurDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Livreurs"]}},"/api/livreurs/{id}/positions":{"get":{"operationId":"LivreursController_historiquePositions","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Livreurs"]}},"/api/commandes":{"post":{"operationId":"CommandesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommandeDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["Commandes"]},"get":{"operationId":"CommandesController_findAll","parameters":[{"name":"statut","required":true,"in":"query","schema":{"type":"string"}},{"name":"livreurId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Commandes"]}},"/api/commandes/mes-commandes":{"get":{"operationId":"CommandesController_mesCommandes","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Commandes"]}},"/api/commandes/{id}":{"get":{"operationId":"CommandesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Commandes"]}},"/api/commandes/{id}/affecter":{"patch":{"operationId":"CommandesController_affecterLivreur","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AffecterLivreurDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Commandes"]}},"/api/commandes/{id}/statut":{"patch":{"operationId":"CommandesController_changerStatut","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangerStatutDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["Commandes"]}},"/api/gps/position":{"post":{"operationId":"GpsController_enregistrerPosition","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnregistrerPositionDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["GPS"]}},"/api/gps/livreurs/dernieres-positions":{"get":{"operationId":"GpsController_dernieresPositions","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["GPS"]}}},"info":{"title":"UNIQUE Livraison Expresse API","description":"API de gestion des clients, commandes, livreurs, GPS et notifications.","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"telephone":{"type":"string","example":"+22670000003"},"motDePasse":{"type":"string","example":"mot-de-passe"}},"required":["telephone","motDePasse"]},"DeviceTokenDto":{"type":"object","properties":{"token":{"type":"object","description":"Jeton FCM; null pour retirer l’appareil","maxLength":4096,"nullable":true}}},"CreateClientDto":{"type":"object","properties":{}},"UpdateClientDto":{"type":"object","properties":{}},"CreateAdresseDto":{"type":"object","properties":{}},"CreateFournisseurDto":{"type":"object","properties":{}},"UpdateFournisseurDto":{"type":"object","properties":{}},"CreateProduitDto":{"type":"object","properties":{}},"CreateLivreurDto":{"type":"object","properties":{}},"UpdateLivreurDto":{"type":"object","properties":{}},"UpdateStatutLivreurDto":{"type":"object","properties":{}},"CreateCommandeDto":{"type":"object","properties":{"clientId":{"type":"string"},"fournisseurId":{"type":"string"},"adresseLivraisonId":{"type":"string"},"descriptionMarchandise":{"type":"string"},"quantite":{"type":"string","example":"2 cartons"},"montantMarchandise":{"type":"number"},"fraisLivraison":{"type":"number"}},"required":["clientId","fournisseurId","adresseLivraisonId","descriptionMarchandise","quantite","fraisLivraison"]},"AffecterLivreurDto":{"type":"object","properties":{"livreurId":{"type":"string"}},"required":["livreurId"]},"ChangerStatutDto":{"type":"object","properties":{"statut":{"type":"string","enum":["NOUVELLE","CONFIRMEE","LIVREUR_AFFECTE","EN_ROUTE_FOURNISSEUR","MARCHANDISE_RECUPEREE","EN_ROUTE_CLIENT","LIVREE","TERMINEE","ANNULEE","ECHEC_LIVRAISON"]},"motifEchec":{"type":"string","enum":["CLIENT_ABSENT","MARCHANDISE_INDISPONIBLE","AUTRE"]}},"required":["statut"]},"EnregistrerPositionDto":{"type":"object","properties":{}}}}}